+++ packages/contracts/src/systems/_NodeDeleteSystem.sol
@@ -15,7 +15,7 @@ contract _NodeDeleteSystem is System {
   constructor(IWorld _world, address _components) System(_world, _components) {}
 
   function execute(bytes memory arguments) public onlyOwner returns (bytes memory) {
-    uint256 index = abi.decode(arguments, (uint256));
+    uint32 index = abi.decode(arguments, (uint32));
     uint256 id = LibNode.getByIndex(components, index);
 
     require(id != 0, "Node: does not exist");
@@ -25,7 +25,7 @@ contract _NodeDeleteSystem is System {
     return "";
   }
 
-  function executeTyped(uint256 index) public onlyOwner returns (bytes memory) {
+  function executeTyped(uint32 index) public onlyOwner returns (bytes memory) {
     return execute(abi.encode(index));
   }
 }
